Move back struct to Cheats and just include header
